+++ NervaOneWalletMiner/Objects/Stats/StatsWallet.cs
@@ -6,8 +6,8 @@ namespace NervaOneWalletMiner.Objects.Stats
     public class StatsWallet
     {
         public Dictionary<uint, Account> Subaddresses { get; set; } = [];
-        public double TotalBalanceLocked { get; set; } = 0.00;
-        public double TotalBalanceUnlocked { get; set; } = 0.00;
+        public decimal TotalBalanceLocked { get; set; } = 0;
+        public decimal TotalBalanceUnlocked { get; set; } = 0;
 
     }
 }
